requirements.txt
appdirs==1.4.0
numpy==1.12.0
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.1.10
six==1.10.0
